+++ src/Scaffolding/Generator/BaseGenerator.php
@@ -19,7 +19,6 @@ namespace Moodle\Composer\Scaffold\Scaffolding\Generator;
 
 use Composer\Composer;
 use Composer\IO\IOInterface;
-use Composer\Util\Filesystem;
 use Moodle\Composer\Scaffold\PackagePathTrait;
 
 /**
@@ -49,4 +48,45 @@ abstract class BaseGenerator
      * @return void
      */
     abstract public function generate(): void;
+
+    /**
+     * Add a path to the .gitignore.
+     *
+     * @param string $path
+     * @param string $description
+     * @param mixed $reason
+     */
+    protected function addGitIgnore(
+        string $path,
+        string $description,
+        ?string $reason = null,
+    ): void {
+        $gitignorepath = $this->getRootPackagePath() . '/.gitignore';
+        $fullentry = "# Ignore {$description}\n";
+        if ($reason) {
+            $fullentry .= "# {$reason}\n";
+        }
+        $entry = '/' . ltrim($path, '/') . "\n";
+        $fullentry .= $entry;
+
+        if (file_exists($gitignorepath)) {
+            $currentcontent = file_get_contents($gitignorepath);
+            if (strpos($currentcontent, $entry) !== false) {
+                $this->io->debug("{$path} is already in .gitignore");
+                return;
+            }
+
+            $this->io->debug("Adding {$path} to .gitignore");
+            $newcontent = $currentcontent . $fullentry;
+            file_put_contents($gitignorepath, $newcontent);
+        } else {
+            if (!file_exists(dirname($gitignorepath) . '/.git')) {
+                $this->io->debug("Not creating .gitignore as project is not under git");
+                return;
+            }
+
+            $this->io->debug("Creating .gitignore and adding {$path}");
+            file_put_contents($gitignorepath, $fullentry);
+        }
+    }
 }